refactor: introduce typed ActClaim model and eliminate redundant JWT parsing
- Add ActClaim class under model/ package to represent the recursive
  'act' claim structure defined in RFC 8693, replacing raw Object/Map
  handling with a typed recursive model

- Centralize unsafe Object-to-Map casting inside ActClaim.fromRawClaim()
  so call sites remain clean and type-safe

- Add ACT constant to TokenExchangeConstants to replace "act" string
  literal usage across the codebase

- Refactor isDelegationRequest() to accept pre-parsed JWTClaimsSet
  instead of parsing the subject token internally, removing redundant
  JWT parsing

- Refactor validateSubjectTokenForDelegation() to accept pre-parsed
  SignedJWT and JWTClaimsSet parameters instead of parsing internally

- Parse subject token once in validateGrant() and pass the result down
  to isDelegationRequest() and validateSubjectTokenForDelegation(),
  eliminating duplicate parsing of the same token

- Remove throws IdentityOAuth2Exception from isDelegationRequest() as
  it no longer performs any operations that can throw
